apps/backend-masterfinder/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "backend" ]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-report html:'../../coverage/apps/backend-masterfinder/html' --cov-report xml:'../../coverage/apps/backend-masterfinder/coverage.xml' --html='../../reports/apps/backend-masterfinder/unittests/html/index.html' --junitxml='../../reports/apps/backend-masterfinder/unittests/junit.xml'"

[tool.poetry]
name = "backend"
version = "1.0.0"
description = "backend"
authors = [ ]
license = 'Proprietary'
readme = 'README.md'

  [[tool.poetry.packages]]
  include = "backend"

  [tool.poetry.dependencies]
  python = ">=3.9,<=3.12.3"
  faker = "^25.5.0"
  fastapi = "^0.111.0"
  sqlalchemy = "^2.0.30"
  pyjwt = "^2.8.0"
  passlib = "^1.7.4"
  bcrypt = "^4.1.3"
  pillow = "^10.3.0"

  [tool.poetry.group.dev.dependencies]
  autopep8 = "2.0.2"
  flake8 = "6.0.0"
  pytest = "7.3.1"
  pytest-sugar = "0.9.7"
  pytest-cov = "4.1.0"
  pytest-html = "3.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
